A dynmaic web page
Create a new dynamic web application which allows users to monitor and control devices in
their smart home. The app is just a prototype so it doesn’t need interface with actual
hardware, but it should provide the following functionality:

Users can add a range of smart home devices to their dashboard. For example:
○ cooling and heating systems,
○ curtains and blinds,
○ lighting, security,
○ audio- visual systems such as TV, radio, music players,
○ kitchen appliances such as fridge-freezer, hob, oven, microwave
● Users can organise their devices by giving them custom names
● Users can monitor the status for each device (Eg. heating - room temperature, on/off,
lighting on or off, curtains open/closed)
● Users can control each of these devices (Eg. turn lighting on or off, increase or
decrease heating temperature etc)